refactor: move outline & author guide components into own files
