+++ src/web/components/guides/mod.rs
@@ -2,4 +2,5 @@ pub mod author;
 pub mod autumn;
 pub mod category;
 pub mod layout;
+pub mod outline;
 pub mod sidebar;
